fix ailment prob to be a float instead of int

+++ src/pokemon/moves/moves.py
@@ -16,4 +16,4 @@ class Moves(BaseModel):
     priority: int
     stat_changes: List[int]
     ailment_name: NVStatus
-    ailment_prob: int
+    ailment_prob: float
